Move components for theme selection and theme editing to a new directory
